+++ driven/storage/adapter.go
@@ -2,7 +2,6 @@ package storage
 
 import (
 	"context"
-	"core-building-block/core/auth"
 	"core-building-block/core/model"
 	"errors"
 	"fmt"
@@ -52,7 +51,7 @@ func (sa *Adapter) Start() error {
 }
 
 //RegisterStorageListener registers a data change listener with the storage adapter
-func (sa *Adapter) RegisterStorageListener(storageListener model.StorageListener) {
+func (sa *Adapter) RegisterStorageListener(storageListener StorageListener) {
 	sa.db.listeners = append(sa.db.listeners, storageListener)
 }
 
@@ -62,9 +61,9 @@ func (sa *Adapter) ReadTODO() error {
 }
 
 //FindAuthConfig finds the auth document from DB by orgID and appID
-func (sa *Adapter) FindAuthConfig(orgID string, appID string, authType string) (*auth.AuthConfig, error) {
+func (sa *Adapter) FindAuthConfig(orgID string, appID string, authType string) (*model.AuthConfig, error) {
 	filter := bson.D{primitive.E{Key: "org_id", Value: orgID}, primitive.E{Key: "app_id", Value: appID}, primitive.E{Key: "type", Value: authType}}
-	var result *auth.AuthConfig
+	var result *model.AuthConfig
 	err := sa.db.authConfigs.FindOne(filter, &result, nil)
 	if err != nil {
 		return nil, err
@@ -76,9 +75,9 @@ func (sa *Adapter) FindAuthConfig(orgID string, appID string, authType string) (
 }
 
 //LoadAuthConfigs finds all auth config documents in the DB
-func (sa *Adapter) LoadAuthConfigs() (*[]auth.AuthConfig, error) {
+func (sa *Adapter) LoadAuthConfigs() (*[]model.AuthConfig, error) {
 	filter := bson.D{}
-	var result []auth.AuthConfig
+	var result []model.AuthConfig
 	err := sa.db.authConfigs.Find(filter, &result, nil)
 	if err != nil {
 		return nil, err
@@ -257,4 +256,14 @@ func abortTransaction(sessionContext mongo.SessionContext) {
 		//TODO - log
 	}
 
+}
+
+type StorageListener interface {
+	OnAuthConfigUpdated()
+}
+
+type DefaultStorageListenerImpl struct {
+}
+
+func (d *DefaultStorageListenerImpl) OnAuthConfigUpdated() {
 }